https://online.sbis.ru/opendoc.html?guid=8d622a59-3886-4034-935e-b5d7b074ed74
модуль MaskType
